+++ stacks-signer/src/v0/signer.rs
@@ -15,17 +15,26 @@
 use std::collections::HashMap;
 use std::fmt::Debug;
 use std::sync::mpsc::Sender;
+#[cfg(any(test, feature = "testing"))]
+use std::sync::LazyLock;
 use std::time::{Duration, Instant};
 
 use blockstack_lib::chainstate::nakamoto::{NakamotoBlock, NakamotoBlockHeader};
 use blockstack_lib::net::api::postblock_proposal::{
-    BlockValidateOk, BlockValidateReject, BlockValidateResponse, TOO_MANY_REQUESTS_STATUS,
+    BlockValidateOk, BlockValidateReject, BlockValidateResponse, ValidateRejectCode,
+    TOO_MANY_REQUESTS_STATUS,
 };
 use blockstack_lib::util_lib::db::Error as DBError;
 use clarity::types::chainstate::StacksPrivateKey;
+#[cfg(any(test, feature = "testing"))]
+use clarity::types::chainstate::StacksPublicKey;
 use clarity::types::{PrivateKey, StacksEpochId};
 use clarity::util::hash::{MerkleHashFunc, Sha512Trunc256Sum};
 use clarity::util::secp256k1::Secp256k1PublicKey;
+#[cfg(any(test, feature = "testing"))]
+use clarity::util::sleep_ms;
+#[cfg(any(test, feature = "testing"))]
+use clarity::util::tests::TestFlag;
 use libsigner::v0::messages::{
     BlockAccepted, BlockRejection, BlockResponse, MessageSlotID, MockProposal, MockSignature,
     RejectReason, RejectReasonPrefix, SignerMessage,
@@ -43,6 +52,12 @@ use crate::config::{SignerConfig, SignerConfigMode};
 use crate::runloop::SignerResult;
 use crate::signerdb::{BlockInfo, BlockState, SignerDb};
 use crate::Signer as SignerTrait;
+
+/// A global variable that can be used to make signers repeat their proposal
+/// response if their public key is in the provided list
+#[cfg(any(test, feature = "testing"))]
+pub static TEST_REPEAT_PROPOSAL_RESPONSE: LazyLock<TestFlag<Vec<StacksPublicKey>>> =
+    LazyLock::new(TestFlag::default);
 
 /// Signer running mode (whether dry-run or real)
 #[derive(Debug)]
@@ -377,6 +392,7 @@ impl Signer {
             ),
         )
     }
+
     /// Create a block rejection response for a block with the given reject code
     pub fn create_block_rejection(
         &self,
@@ -397,6 +413,7 @@ impl Signer {
             ),
         )
     }
+
     /// Check if block should be rejected based on sortition state
     /// Will return a BlockResponse::Rejection if the block is invalid, none otherwise.
     fn check_block_against_sortition_state(
@@ -464,6 +481,45 @@ impl Signer {
         }
     }
 
+    /// The actual `send_block_response` implementation. Declared so that we do
+    /// not need to duplicate in testing.
+    fn impl_send_block_response(&mut self, block_response: BlockResponse) {
+        let res = self
+            .stackerdb
+            .send_message_with_retry::<SignerMessage>(block_response.clone().into());
+        match res {
+            Err(e) => warn!("{self}: Failed to send block rejection to stacker-db: {e:?}"),
+            Ok(ack) if !ack.accepted => warn!(
+                "{self}: Block rejection not accepted by stacker-db: {:?}",
+                ack.reason
+            ),
+            Ok(_) => debug!("{self}: Block rejection accepted by stacker-db"),
+        }
+    }
+
+    #[cfg(any(test, feature = "testing"))]
+    fn send_block_response(&mut self, block_response: BlockResponse) {
+        const NUM_REPEATS: usize = 1;
+        let mut count = 0;
+        let public_keys = TEST_REPEAT_PROPOSAL_RESPONSE.get();
+        if !public_keys.contains(
+            &stacks_common::types::chainstate::StacksPublicKey::from_private(&self.private_key),
+        ) {
+            count = NUM_REPEATS;
+        }
+        while count <= NUM_REPEATS {
+            self.impl_send_block_response(block_response.clone());
+
+            count += 1;
+            sleep_ms(1000);
+        }
+    }
+
+    #[cfg(not(any(test, feature = "testing")))]
+    fn send_block_response(&mut self, block_response: BlockResponse) {
+        self.impl_send_block_response(block_response)
+    }
+
     /// Handle block proposal messages submitted to signers stackerdb
     fn handle_block_proposal(
         &mut self,
@@ -503,33 +559,19 @@ impl Signer {
         // TODO: should add a check to ignore an old burn block height if we know its outdated. Would require us to store the burn block height we last saw on the side.
         //  the signer needs to be able to determine whether or not the block they're about to sign would conflict with an already-signed Stacks block
         let signer_signature_hash = block_proposal.block.header.signer_signature_hash();
-        if let Some(block_info) = self.block_lookup_by_reward_cycle(&signer_signature_hash) {
-            let Some(block_response) = self.determine_response(&block_info) else {
-                // We are still waiting for a response for this block. Do nothing.
-                debug!("{self}: Received a block proposal for a block we are already validating.";
-                    "signer_sighash" => %signer_signature_hash,
-                    "block_id" => %block_proposal.block.block_id()
-                );
-                return;
-            };
-            // Submit a proposal response to the .signers contract for miners
-            debug!("{self}: Broadcasting a block response to stacks node: {block_response:?}");
-            let accepted = matches!(block_response, BlockResponse::Accepted(..));
-            match self
-                .stackerdb
-                .send_message_with_retry::<SignerMessage>(block_response.into())
-            {
-                Ok(_) => {
-                    crate::monitoring::actions::increment_block_responses_sent(accepted);
-                    crate::monitoring::actions::record_block_response_latency(
-                        &block_proposal.block,
-                    );
-                }
-                Err(e) => {
-                    warn!("{self}: Failed to send block response to stacker-db: {e:?}",);
-                }
-            }
-            return;
+        let prior_evaluation = self
+            .block_lookup_by_reward_cycle(&signer_signature_hash)
+            .and_then(|block_info| if should_reevaluate_block(&block_info) {
+                debug!("Received a proposal for this block before, but our rejection reason allows us to reconsider";
+                    "reject_reason" => ?block_info.reject_reason);
+                None
+            } else {
+                Some(block_info)
+            });
+
+        // we previously considered this proposal, handle the status here
+        if let Some(block_info) = prior_evaluation {
+            return self.handle_prior_proposal_eval(&block_info);
         }
 
         info!(
@@ -575,18 +617,7 @@ impl Signer {
         if let Some(block_response) = block_response {
             // We know proposal is invalid. Send rejection message, do not do further validation and do not store it.
             debug!("{self}: Broadcasting a block response to stacks node: {block_response:?}");
-            let res = self
-                .stackerdb
-                .send_message_with_retry::<SignerMessage>(block_response.into());
-
-            match res {
-                Err(e) => warn!("{self}: Failed to send block rejection to stacker-db: {e:?}"),
-                Ok(ack) if !ack.accepted => warn!(
-                    "{self}: Block rejection not accepted by stacker-db: {:?}",
-                    ack.reason
-                ),
-                Ok(_) => debug!("{self}: Block rejection accepted by stacker-db"),
-            }
+            self.send_block_response(block_response);
         } else {
             // Just in case check if the last block validation submission timed out.
             self.check_submitted_block_proposal();
@@ -622,6 +653,32 @@ impl Signer {
             self.signer_db
                 .insert_block(&block_info)
                 .unwrap_or_else(|e| self.handle_insert_block_error(e));
+        }
+    }
+
+    fn handle_prior_proposal_eval(&mut self, block_info: &BlockInfo) {
+        let Some(block_response) = self.determine_response(block_info) else {
+            // We are still waiting for a response for this block. Do nothing.
+            debug!(
+                "{self}: Received a block proposal for a block we are already validating.";
+                "signer_sighash" => %block_info.signer_signature_hash(),
+                "block_id" => %block_info.block.block_id()
+            );
+            return;
+        };
+
+        // Submit a proposal response to the .signers contract for miners
+        debug!("{self}: Broadcasting a block response to stacks node: {block_response:?}");
+
+        let accepted = matches!(block_response, BlockResponse::Accepted(..));
+        if let Err(e) = self
+            .stackerdb
+            .send_message_with_retry::<SignerMessage>(block_response.into())
+        {
+            warn!("{self}: Failed to send block response to stacker-db: {e:?}");
+        } else {
+            crate::monitoring::actions::increment_block_responses_sent(accepted);
+            crate::monitoring::actions::record_block_response_latency(&block_info.block);
         }
     }
 
@@ -844,6 +901,8 @@ impl Signer {
                 false,
             ),
         );
+
+        block_info.reject_reason = Some(block_rejection.response_data.reject_reason.clone());
         self.signer_db
             .insert_block(&block_info)
             .unwrap_or_else(|e| self.handle_insert_block_error(e));
@@ -973,6 +1032,7 @@ impl Signer {
             ),
             &block_info.block,
         );
+        block_info.reject_reason = Some(rejection.get_response_data().reject_reason.clone());
         if let Err(e) = block_info.mark_locally_rejected() {
             if !block_info.has_reached_consensus() {
                 warn!("{self}: Failed to mark block as locally rejected: {e:?}");
@@ -993,6 +1053,7 @@ impl Signer {
             ),
             Ok(_) => debug!("{self}: Block rejection accepted by stacker-db"),
         }
+
         self.signer_db
             .insert_block(&block_info)
             .unwrap_or_else(|e| self.handle_insert_block_error(e));
@@ -1086,6 +1147,7 @@ impl Signer {
         ) {
             warn!("{self}: Failed to save block rejection signature: {e:?}",);
         }
+        block_info.reject_reason = Some(rejection.response_data.reject_reason.clone());
 
         // do we have enough signatures to mark a block a globally rejected?
         // i.e. is (set-size) - (threshold) + 1 reached.
@@ -1364,5 +1426,35 @@ impl Signer {
         } else {
             None
         }
+    }
+}
+
+/// Determine if a block should be re-evaluated based on its rejection reason˝
+fn should_reevaluate_block(block_info: &BlockInfo) -> bool {
+    if let Some(reject_reason) = &block_info.reject_reason {
+        match reject_reason {
+            RejectReason::ValidationFailed(ValidateRejectCode::UnknownParent)
+            | RejectReason::NoSortitionView
+            | RejectReason::ConnectivityIssues(_)
+            | RejectReason::TestingDirective
+            | RejectReason::InvalidTenureExtend
+            | RejectReason::NotRejected
+            | RejectReason::Unknown(_) => true,
+            RejectReason::ValidationFailed(_)
+            | RejectReason::RejectedInPriorRound
+            | RejectReason::SortitionViewMismatch
+            | RejectReason::ReorgNotAllowed
+            | RejectReason::InvalidBitvec
+            | RejectReason::PubkeyHashMismatch
+            | RejectReason::InvalidMiner
+            | RejectReason::NotLatestSortitionWinner
+            | RejectReason::InvalidParentBlock
+            | RejectReason::DuplicateBlockFound => {
+                // No need to re-validate these types of rejections.
+                false
+            }
+        }
+    } else {
+        false
     }
 }